+++ src/scripts/productivity.sh
@@ -1,5 +1,9 @@
 # Install taskwarrior
-sudo dnf install task -y
+if [[ -f "/usr/bin/task" ]]; then
+	echo "Taskwarrior is already installed."
+else
+	sudo dnf install task -y
+fi
 
 # Taskwarrior config
 cat "$(pwd)/src/artifacts/taskwarrior/taskrcUpdates.txt" >> ~/.taskrc
@@ -20,10 +24,11 @@ cp "$(pwd)/src/artifacts/taskwarrior/themes/" ~/.task/themes/
 # Install Simplenote and Todoist
 flatpakApps=("com.simplenote.Simplenote" "com.todoist.Todoist")
 for flatpakApp in ${flatpakApps[@]}; do
-	# TODO: Path to flatpak apps is either /var/lib/flatpak or ~/.local/share/flatpak
-	if [[ -d "path/to/$flatpakApp" ]]; then
-		echo "$flatpakApp is already installed."
+	if [[ -d "/var/lib/flatpak/app/$flatpakApp" ]]; then
+		echo "$flatpak is already installed."
+	elif [[ -d "~/.local/share/flatpak/app/$flatpakApp"]]; then
+		echo "$flatpak is already installed."
 	else
-		brew install --cask "$flatpakApp"
+		sudo dnf install "$flatpakApp" -y
 	fi
 done
