+++ src/scripts/security.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 
 # Install packages for YubiKeys
-sudo dnf install pam pam-u2f pamu2fcfg -y
+if [[ -f "/usr/bin/pamu2fcfg" ]]; then
+    echo "pam modules are already installed."
+else
+    sudo dnf install pam pam-u2f pamu2fcfg -y
+fi
 
 # Create local directory for key registration
 mkdir -p ~/.config/yubico
@@ -37,15 +41,23 @@ sudo chmod 644 /etc/yubico/u2f_keys
 # TODO: Add python script to update /etc/pam.d/sudo to add: auth sufficient pam_u2f.so authfile=/etc/yubico/u2f_keys
 
 # Install and Enable Firewall
-sudo dnf install ufw -y
+if [[ -f "/usr/sbin/ufw" ]]; then
+    echo "Firewall is already installed."
+else
+    sudo dnf install ufw -y
+fi
 sudo ufw enable
 
 ## Install 1Password
-sudo rpm --import https://downloads.1password.com/linux/keys/1password.asc
-sudo sh -c 'echo -e "[1password]\nname=1Password Stable Channel\nbaseurl=https://downloads.1password.com/linux/rpm/stable/\$basearch\nenabled=1\ngpgcheck=1\nrepo_gpgcheck=1\ngpgkey=\"https://downloads.1password.com/linux/keys/1password.asc\"" > /etc/yum.repos.d/1password.repo'
-sudo dnf install 1password -y
+if [[ -f "/usr/bin/1password" ]]; then
+    echo "1Password is already installed."
+else
+    sudo rpm --import https://downloads.1password.com/linux/keys/1password.asc
+    sudo sh -c 'echo -e "[1password]\nname=1Password Stable Channel\nbaseurl=https://downloads.1password.com/linux/rpm/stable/\$basearch\nenabled=1\ngpgcheck=1\nrepo_gpgcheck=1\ngpgkey=\"https://downloads.1password.com/linux/keys/1password.asc\"" > /etc/yum.repos.d/1password.repo'
+    sudo dnf install 1password -y
+fi
 
-# Install Proton VPN
+# TODO: Install Proton VPN
 
 # Download and Install RPM Package
 # TODO: Download RPM Package from https://protonvpn.com/download/protonvpn-stable-release-1.0.1-1.noarch.rpm
@@ -58,8 +70,16 @@ sudo dnf install python3-pip -y
 sudo dnf install --user 'dnspython>=1.16.0' -y
 
 # Install Clam AV
-sudo dnf upgrade --refresh -y
-sudo dnf install clamav clamd clamav-update -y
+if [[ -f "/usr/bin/clamscan" ]]; then
+    echo "Clam Anti-Virus is already installed."
+else  
+    sudo dnf upgrade --refresh -y
+    sudo dnf install clamav clamd clamav-update -y
+fi
 
 # Install nmap
-sudo dnf install nmap -y
+if [[ -f "/usr/bin/clamscan" ]]; then
+    echo "nmap is already installed."
+else
+    sudo dnf install nmap -y
+fi
